Added `sender`/`receiver` modules.
- Shifted share logic to `core` module.
